+++ components/Footer.vue
@@ -115,7 +115,7 @@
 
 <script lang="ts">
 import { Component, Vue } from 'nuxt-property-decorator'
-import i18n from '@/i18n'
+import i18n from '@/utils/i18n/i18n'
 import { TranslateResult } from 'vue-i18n/types'
 
 interface Menu {
